+++ thread_raytracing/Math/Container/Block.hpp
@@ -11,7 +11,7 @@ struct block
 	block();
 	block(int size);
 
-	std::vector<vec3<float>> chunk;
+	std::vector<vec3<double>> chunk;
 };
