+++ thread_raytracing/Math/Ray/Ray.cpp
@@ -6,15 +6,15 @@ namespace math
 		:direction{}, origin{}
 	{}
 
-	ray::ray(vec3<float> direction)
+	ray::ray(vec3<double> direction)
 		: direction{direction}, origin{}
 	{}
 
-	ray::ray(vec3<float> origin, vec3<float> direction)
+	ray::ray(vec3<double> origin, vec3<double> direction)
 		: origin{origin}, direction{direction}
 	{}
 
-	vec3<float> ray::at(float t) const
+	vec3<double> ray::at(double t) const
 	{
 		return origin + direction * t;
 	}
